requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-telegram-bot==20.7
sqlalchemy==2.0.27
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
typer==0.9.0
rich==13.7.0
loguru==0.7.2
httpx==0.25.2
aiogram==3.7.0

# dev tools
black==24.10.0
isort==5.13.2
mypy==1.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# other
aiosqlite==0.19.0
apscheduler==3.10.4
redis==5.3.0
alembic>=1.12.1
python-multipart==0.0.6
pyngrok  # можешь указать версию, например: pyngrok==5.3.0
